test: add customer balance tests with various order and payment scenarios
